% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-mp.R
\name{all_mp_fns}
\alias{all_mp_fns}
\title{Get All Meta-Property Functions}
\usage{
all_mp_fns()
}
\value{
A named list of meta-property functions.
}
\description{
This function returns a named list of all meta-property functions:
\itemize{
\item "T": \code{n_glycan_type()}: type of the glycan
\item "B": \code{has_bisecting()}: whether the glycan has a bisecting GlcNAc
\item "nA": \code{n_antennae()}: number of antennae
\item "nFc": \code{n_core_fuc()}: number of core fucoses
\item "nFa": \code{n_arm_fuc()}: number of arm fucoses
\item "nG": \code{n_gal()}: number of galactoses
\item "nGt": \code{n_terminal_gal()}: number of terminal galactoses
\item "nS": \code{n_sia()}: number of sialic acids
\item "nM": \code{n_man()}: number of mannoses
}
}
