% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{all_traits}
\alias{all_traits}
\title{Get All Derived Traits}
\usage{
all_traits()
}
\value{
A named list of derived traits.
}
\description{
This function returns a named list of all derived traits.
Compared to \code{\link[=basic_traits]{basic_traits()}}, this function includes derived traits with more
detailed \code{within} conditions.
}
\details{
The explanations of the derived traits are as follows:
\itemize{
\item \code{A1F}: Proportion of fucosylated glycans within mono-antennary glycans
\item \code{A2F}: Proportion of fucosylated glycans within bi-antennary glycans
\item \code{A3F}: Proportion of fucosylated glycans within tri-antennary glycans
\item \code{A4F}: Proportion of fucosylated glycans within tetra-antennary glycans
\item \code{A1Fc}: Proportion of core-fucosylated glycans within mono-antennary glycans
\item \code{A2Fc}: Proportion of core-fucosylated glycans within bi-antennary glycans
\item \code{A3Fc}: Proportion of core-fucosylated glycans within tri-antennary glycans
\item \code{A4Fc}: Proportion of core-fucosylated glycans within tetra-antennary glycans
\item \code{A1Fa}: Proportion of arm-fucosylated glycans within mono-antennary glycans
\item \code{A2Fa}: Proportion of arm-fucosylated glycans within bi-antennary glycans
\item \code{A3Fa}: Proportion of arm-fucosylated glycans within tri-antennary glycans
\item \code{A4Fa}: Proportion of arm-fucosylated glycans within tetra-antennary glycans
\item \code{A1SFa}: Proportion of arm-fucosylated glycans within sialylated mono-antennary glycans
\item \code{A2SFa}: Proportion of arm-fucosylated glycans within sialylated bi-antennary glycans
\item \code{A3SFa}: Proportion of arm-fucosylated glycans within sialylated tri-antennary glycans
\item \code{A4SFa}: Proportion of arm-fucosylated glycans within sialylated tetra-antennary glycans
\item \code{A1S0Fa}: Proportion of arm-fucosylated glycans within asialylated mono-antennary glycans
\item \code{A2S0Fa}: Proportion of arm-fucosylated glycans within asialylated bi-antennary glycans
\item \code{A3S0Fa}: Proportion of arm-fucosylated glycans within asialylated tri-antennary glycans
\item \code{A4S0Fa}: Proportion of arm-fucosylated glycans within asialylated tetra-antennary glycans
\item \code{A1B}: Proportion of bisecting glycans within mono-antennary glycans
\item \code{A2B}: Proportion of bisecting glycans within bi-antennary glycans
\item \code{A3B}: Proportion of bisecting glycans within tri-antennary glycans
\item \code{A4B}: Proportion of bisecting glycans within tetra-antennary glycans
\item \code{A1FcB}: Proportion of bisecting glycans within core-fucosylated mono-antennary glycans
\item \code{A2FcB}: Proportion of bisecting glycans within core-fucosylated bi-antennary glycans
\item \code{A3FcB}: Proportion of bisecting glycans within core-fucosylated tri-antennary glycans
\item \code{A4FcB}: Proportion of bisecting glycans within core-fucosylated tetra-antennary glycans
\item \code{A1Fc0B}: Proportion of bisecting glycans within a-core-fucosylated mono-antennary glycans
\item \code{A2Fc0B}: Proportion of bisecting glycans within a-core-fucosylated bi-antennary glycans
\item \code{A3Fc0B}: Proportion of bisecting glycans within a-core-fucosylated tri-antennary glycans
\item \code{A4Fc0B}: Proportion of bisecting glycans within a-core-fucosylated tetra-antennary glycans
\item \code{A1G}: Average degree of galactosylation per antenna within mono-antennary glycans
\item \code{A2G}: Average degree of galactosylation per antenna within bi-antennary glycans
\item \code{A3G}: Average degree of galactosylation per antenna within tri-antennary glycans
\item \code{A4G}: Average degree of galactosylation per antenna within tetra-antennary glycans
\item \code{A1Gt}: Average degree of terminal galactose per antenna within mono-antennary glycans
\item \code{A2Gt}: Average degree of terminal galactose per antenna within bi-antennary glycans
\item \code{A3Gt}: Average degree of terminal galactose per antenna within tri-antennary glycans
\item \code{A4Gt}: Average degree of terminal galactose per antenna within tetra-antennary glycans
\item \code{A1S}: Average degree of sialylation per antenna within mono-antennary glycans
\item \code{A2S}: Average degree of sialylation per antenna within bi-antennary glycans
\item \code{A3S}: Average degree of sialylation per antenna within tri-antennary glycans
\item \code{A4S}: Average degree of sialylation per antenna within tetra-antennary glycans
\item \code{A1GS}: Average degree of sialylation per galactose within mono-antennary glycans
\item \code{A2GS}: Average degree of sialylation per galactose within bi-antennary glycans
\item \code{A3GS}: Average degree of sialylation per galactose within tri-antennary glycans
\item \code{A4GS}: Average degree of sialylation per galactose within tetra-antennary glycans
}
}
\examples{
all_traits()

}
