% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{basic_traits}
\alias{basic_traits}
\title{Get Basic Derived Traits}
\usage{
basic_traits()
}
\value{
A named list of derived traits.
}
\description{
These derived traits are the most basic and commonly used derived traits.
They describe global properties of a glycome including the type of glycans,
fucosylation level, sialylation level, galactosylation level,
and branching level.
}
\details{
The explanations of the derived traits are as follows:
\itemize{
\item \code{TM}: Proportion of highmannose glycans
\item \code{TH}: Proportion of hybrid glycans
\item \code{TC}: Proportion of complex glycans
\item \code{MM}: Average number of mannoses within highmannose glycans
\item \code{CA2}: Proportion of bi-antennary glycans within complex glycans
\item \code{CA3}: Proportion of tri-antennary glycans within complex glycans
\item \code{CA4}: Proportion of tetra-antennary glycans within complex glycans
\item \code{TF}: Proportion of fucosylated glycans
\item \code{TFc}: Proportion of core-fucosylated glycans
\item \code{TFa}: Proportion of arm-fucosylated glycans
\item \code{TB}: Proportion of glycans with bisecting GlcNAc
\item \code{SG}: Average degree of sialylation per galactose
\item \code{GA}: Average degree of galactosylation per antenna
\item \code{TS}: Proportion of sialylated glycans
}
}
\examples{
basic_traits()

}
