% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive-traits.R
\name{derive_traits}
\alias{derive_traits}
\title{Calculate Derived Traits}
\usage{
derive_traits(exp, trait_fns = NULL, mp_fns = NULL, mp_cols = NULL)
}
\arguments{
\item{exp}{A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} or \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} object.
Before using this function,
you should preprocess the data using the \code{glyclean} package.
For glycoproteomics data, the data should be aggregated to the
"gfs" (glycoforms with structures) level using \code{glyclean::aggregate()}.
Also, please make sure that the \code{glycan_structure} column is present in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::rowData()}},
as not all glycoproteomics identification softwares provide this information.
"glycan_structure" can be a \code{glyrepr::glycan_structure()} vector,
or a character vector of glycan structure strings supported by \code{glyparse::auto_parse()}.}

\item{trait_fns}{A named list of derived trait functions created by trait factories.
Names of the list are the names of the derived traits.
Default is \code{NULL}, which means all derived traits in \code{\link[=traits_basic]{traits_basic()}} are calculated.}

\item{mp_fns}{A named list of meta-property functions.
This parameter is useful if your trait functions use custom meta-properties
other than those in \code{\link[=all_mp_fns]{all_mp_fns()}}.
Default is \code{NULL}, which means all meta-properties in \code{\link[=all_mp_fns]{all_mp_fns()}} are used.}

\item{mp_cols}{A character vector of column names in \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::rowData()}}
to use as meta-properties.
If names are provided, they will be used as names of the meta-properties,
otherwise the column names will be used.
When \code{mp_cols} is specified, the selected columns overwrite meta-properties introduced by \code{mp_fns}
with the same names, including built-in meta-properties.
Default is \code{NULL}, which means all columns in \code{rowData()} are available as
meta-properties by their existing names. In this default mode,
meta-properties introduced by \code{mp_fns} take precedence over \code{rowData()}
columns with the same names.}
}
\value{
New \code{GlycomicSE} and \code{GlycoproteomicSE} inputs return a plain
\code{SummarizedExperiment}; compatible legacy glyexp inputs preserve their
legacy container type.
Instead of "quantification of each glycan on each glycosite in each sample",
its assay contains "the value of each derived trait on each glycosite in each
sample", with the following columns in \code{rowData()}:
\itemize{
\item \code{trait}: derived trait name
\item \code{explanation}: a concise English explanation of the trait
}

For glycoproteomics data, with additional columns:
\itemize{
\item \code{protein}: protein ID
\item \code{protein_site}: the glycosite position on the protein
}

Other columns in \code{rowData()} (e.g. \code{gene}) are retained if they have "many-to-one"
relationship with glycosites (unique combinations of \code{protein}, \code{protein_site}).
That is, each glycosite cannot have multiple values for these columns.
\code{gene} is a common example, as a glycosite can only be associate with one gene.
Descriptions about glycans are not such a column, as a glycosite can have multiple glycans,
thus having multiple descriptions.
Columns not having this relationship with glycosites will be dropped.
Don't worry if you cannot understand this logic,
as long as you know that this function will try its best to preserve useful information.

\code{colData()} and \code{metadata()} are not modified, except that the \code{exp_type}
field of \code{metadata()} is set to "traitomics" for glycomics data and
"traitproteomics" for glycoproteomics data.
}
\description{
This function calculates derived traits from a \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} or
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} object.
For glycomics data, it calculates the derived traits directly.
For glycoproteomics data, each glycosite is treated as a separate glycome,
and derived traits are calculated in a site-specific manner.
}
\examples{
library(glyexp)
library(SummarizedExperiment)
library(glyclean)

gp_se <- real_experiment |>
  auto_clean() |>
  slice_sample_row(n = 100)
trait_se <- derive_traits(gp_se)
rowData(trait_se)
assay(trait_se)[1:5, 1:5]
colData(trait_se)

# By default, only basic traits are calculated
names(traits_basic())

# You can calculate detailed traits in `traits_detailed()`
more_trait_se <- derive_traits(gp_se, trait_fns = traits_detailed())
more_trait_se

}
\seealso{
\code{\link[=traits_basic]{traits_basic()}}, \code{\link[=traits_detailed]{traits_detailed()}}
}
