% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-properties.R
\name{get_meta_properties}
\alias{get_meta_properties}
\title{Get Meta-Properties of Glycans}
\usage{
get_meta_properties(glycans, mp_fns = NULL)
}
\arguments{
\item{glycans}{A \code{glyrepr::glycan_structure()} vector, or a character vector of glycan structure strings
supported by \code{glyparse::auto_parse()}.}

\item{mp_fns}{A named list of meta-property functions.
Names of the list are the names of the meta-properties. Default is \code{\link[=all_mp_fns]{all_mp_fns()}}.
A meta-property function should takes a \code{glyrepr::glycan_structure()} vector,
and returns a vector of the meta-property values.
purrr-style lambda functions are supported.}
}
\value{
A tibble with the meta-properties. Column names are the names of the meta-properties.
}
\description{
This function calculates the meta-properties of the given glycans.
Meta-properties are properties describing certain structural characteristics of glycans.
For example, the number of antennae, the number of core fucoses, etc.
}
\examples{
glycans <- c(
  "Man(a1-3)[Man(a1-6)]Man(b1-4)GlcNAc(b1-4)GlcNAc(b1-",
  "Fuc(a1-3)GlcNAc(b1-2)Man(a1-3)[GlcNAc(b1-2)Man(a1-6)]Man(b1-4)GlcNAc(b1-4)GlcNAc(?1-",
  "GlcNAc(b1-2)Man(a1-3)[Man(a1-6)]Man(b1-4)GlcNAc(b1-4)GlcNAc(?1-"
)

# Use default meta-property functions
get_meta_properties(glycans)

# Use custom meta-property functions
fns <- list(
  nN = ~ glyrepr::count_mono(.x, "HexNAc"),  # purrr-style lambda function
  nH = ~ glyrepr::count_mono(.x, "Hex")
)
get_meta_properties(glycans, fns)

}
\seealso{
\code{\link[=all_mp_fns]{all_mp_fns()}}
}
