% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-package.R
\docType{package}
\name{glydet-package}
\alias{glydet}
\alias{glydet-package}
\title{glydet: Describe Glycosylation Structural Properties in a Site Specific Manner}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate and analyze glycosylation derived traits for site-specific glycoproteomics data. Derived traits are summary features that aggregate individual glycan abundances into biologically meaningful groups (e.g., galactosylation, sialylation, fucosylation, branching). This package provides functions to compute well-established derived traits from the literature, and features a domain-specific language for defining custom derived traits tailored to specific research needs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glydet/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
