% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-factory.R
\name{prop}
\alias{prop}
\title{Create a Propotion Trait}
\usage{
prop(cond, within = NULL, na_action = "keep")
}
\arguments{
\item{cond}{Condition to use for defining the smaller group.
An expression that evaluates to a logical vector.
The names of all built-in meta-properties (see \code{\link[=all_mp_fns]{all_mp_fns()}}) can be used in the expression.}

\item{within}{Condition to use for defining the larger group,
with the same format as \code{cond}.
If \code{NULL} (default), all glycans are used as the larger group.}

\item{na_action}{How to handle missing values.
\itemize{
\item "keep" (default): keep the missing values as NA.
\item "zero": set the missing values to 0.
}}
}
\value{
A derived trait function.
}
\description{
A proportion trait is the proportion of certain group of glycans within a larger group of glycans.
For example, the proportion of sialylated glycans within all glycans,
or the proportion of tetra-antennary glycans within all complex glycans.
This type of traits is the most common type of glycan derived traits.
}
\examples{
# Proportion of core-fucosylated glycans within all glycans
prop(nFc > 0)

# Proportion of complex glycans within all glycans
prop(T == "complex")

# Proportion of sialylated and fucosylated glycans within all glycans
prop(nS > 0 & nFc > 0)

# Proportion of tetra-antennary glycans within all complex glycans
prop(nFc > 0, within = (T == "complex"))

}
