% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-factory.R
\name{ratio}
\alias{ratio}
\title{Create a Ratio Trait}
\usage{
ratio(num_cond, denom_cond, within = NULL, na_action = "keep")
}
\arguments{
\item{num_cond}{Condition to use for defining the numerator.
An expression that evaluates to a logical vector.
The names of all built-in meta-properties (see \code{\link[=all_mp_fns]{all_mp_fns()}}) and custom meta-properties
can be used in the expression.}

\item{denom_cond}{Condition to use for defining the denominator. Same format as \code{num_cond}.}

\item{within}{Condition to set a restriction for the glycans. Same format as \code{num_cond}.}

\item{na_action}{How to handle missing values.
\itemize{
\item "keep" (default): keep the missing values as NA.
\item "zero": set the missing values to 0.
}}
}
\value{
A derived trait function.
}
\description{
A ratio trait is the ratio of total abundance of two groups of glycans.
For example, the ratio of complex glycans and hybrid glycans,
or the ratio of bisecting and unbisecting glycans.
}
\details{
You can use \code{ratio()} to create ratio trait easily.

For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Ratio of complex glycans and hybrid glycans
ratio(T == "complex", T == "hybrid")

# Ratio of bisecting and unbisecting glycans
ratio(B, !B)

# Ratio of core-fucosylated and non-core-fucosylated glycans within complex glycans
ratio(nFc > 0 & T == "complex", nFc == 0 & T == "complex")

# The above example can be simplified as:
ratio(nFc > 0, nFc == 0, within = (T == "complex"))  # more readable
}\if{html}{\out{</div>}}

Note that the last example uses \code{&} for logical AND.
Actually, you can use any logical operator in the expression in R (e.g., \code{|}, \code{!}, etc.).

\code{prop()} is a special case of \code{ratio()},
i.e., \code{prop(cond, within)} is equivalent to \code{ratio(cond & within, within)}.
We recommend using \code{prop()} instead of \code{ratio()} for clarity if possible.
}
