% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-factory.R
\name{total}
\alias{total}
\title{Create a Total Abundance Trait}
\usage{
total(cond)
}
\arguments{
\item{cond}{Condition to use for defining the group of glycans.
An expression that evaluates to a logical vector.
The names of all built-in meta-properties (see \code{\link[=all_mp_fns]{all_mp_fns()}}) and custom meta-properties
can be used in the expression.}
}
\value{
A derived trait function.
}
\description{
A total abundance trait is the total abundance of a group of glycans.
For example, the total abundance of all complex glycans,
or the total abundance of all tetra-antennary glycans.
}
\section{How to use}{


You can use \code{total()} to create total abundance trait easily.

For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Total abundance of all complex glycans
total(Tp == "complex")

# Total abundance of all tetra-antennary glycans
total(nA == 4)
}\if{html}{\out{</div>}}
}

\examples{
# Total abundance of all complex glycans
total(Tp == "complex")

# Total abundance of all tetra-antennary glycans
total(nA == 4)

}
