% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{traits_clerc_2018}
\alias{traits_clerc_2018}
\title{Get Traits in Clerc et al. 2018}
\usage{
traits_clerc_2018(sia_link = FALSE)
}
\arguments{
\item{sia_link}{A boolean indicating whether to include sialic acid linkage traits. Default is \code{FALSE}.}
}
\value{
A named list of derived traits.
}
\description{
These traits are the ones used by Clerc et al. 2018 (https://doi.org/10.1053/j.gastro.2018.05.030).
We generally don't recommend using these traits because they are either redundant
or missing some important traits.
We include this function because this paper is very influential in the field.
}
\section{Usage of sialic acid linkage traits}{
To use these sialic acid linkage traits,
\code{var_info} of the input \code{glyexp::experiment()} must have the following columns:
\itemize{
\item \code{nE}: Number of a2,6-linked sialic acids
\item \code{nL}: Number of a2,3-linked sialic acids
}

Note that you have to add these two columns even if the \code{glycan_structure} column has intact linkages.
This is because by convention all traits work with glycan structures with "basic" structure levels
(i.e., with generic monosaccharides like "Hex" and "HexNAc" and no linkages specified).
}

\examples{
traits_clerc_2018()[1:5]
}
