% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{traits_detailed}
\alias{traits_detailed}
\alias{all_traits}
\title{Get Detailed Derived Traits}
\usage{
traits_detailed(sia_link = FALSE)

all_traits(sia_link = FALSE)
}
\arguments{
\item{sia_link}{A boolean indicating whether to include sialic acid linkage traits. Default is \code{FALSE}.}
}
\value{
A named list of derived traits.
}
\description{
This function returns a named list of detailed derived traits.
Compared to \code{\link[=traits_basic]{traits_basic()}}, this function includes derived traits with more
detailed \code{within} conditions.
}
\details{
The explanations of the derived traits are as follows:
\itemize{
\item \code{A1F}: Proportion of fucosylated glycans within mono-antennary glycans
\item \code{A2F}: Proportion of fucosylated glycans within bi-antennary glycans
\item \code{A3F}: Proportion of fucosylated glycans within tri-antennary glycans
\item \code{A4F}: Proportion of fucosylated glycans within tetra-antennary glycans
\item \code{A1Fc}: Proportion of core-fucosylated glycans within mono-antennary glycans
\item \code{A2Fc}: Proportion of core-fucosylated glycans within bi-antennary glycans
\item \code{A3Fc}: Proportion of core-fucosylated glycans within tri-antennary glycans
\item \code{A4Fc}: Proportion of core-fucosylated glycans within tetra-antennary glycans
\item \code{A1Fa}: Proportion of arm-fucosylated glycans within mono-antennary glycans
\item \code{A2Fa}: Proportion of arm-fucosylated glycans within bi-antennary glycans
\item \code{A3Fa}: Proportion of arm-fucosylated glycans within tri-antennary glycans
\item \code{A4Fa}: Proportion of arm-fucosylated glycans within tetra-antennary glycans
\item \code{A1SFa}: Proportion of arm-fucosylated glycans within sialylated mono-antennary glycans
\item \code{A2SFa}: Proportion of arm-fucosylated glycans within sialylated bi-antennary glycans
\item \code{A3SFa}: Proportion of arm-fucosylated glycans within sialylated tri-antennary glycans
\item \code{A4SFa}: Proportion of arm-fucosylated glycans within sialylated tetra-antennary glycans
\item \code{A1S0Fa}: Proportion of arm-fucosylated glycans within asialylated mono-antennary glycans
\item \code{A2S0Fa}: Proportion of arm-fucosylated glycans within asialylated bi-antennary glycans
\item \code{A3S0Fa}: Proportion of arm-fucosylated glycans within asialylated tri-antennary glycans
\item \code{A4S0Fa}: Proportion of arm-fucosylated glycans within asialylated tetra-antennary glycans
\item \code{A1B}: Proportion of bisecting glycans within mono-antennary glycans
\item \code{A2B}: Proportion of bisecting glycans within bi-antennary glycans
\item \code{A3B}: Proportion of bisecting glycans within tri-antennary glycans
\item \code{A4B}: Proportion of bisecting glycans within tetra-antennary glycans
\item \code{A1FcB}: Proportion of bisecting glycans within core-fucosylated mono-antennary glycans
\item \code{A2FcB}: Proportion of bisecting glycans within core-fucosylated bi-antennary glycans
\item \code{A3FcB}: Proportion of bisecting glycans within core-fucosylated tri-antennary glycans
\item \code{A4FcB}: Proportion of bisecting glycans within core-fucosylated tetra-antennary glycans
\item \code{A1Fc0B}: Proportion of bisecting glycans within a-core-fucosylated mono-antennary glycans
\item \code{A2Fc0B}: Proportion of bisecting glycans within a-core-fucosylated bi-antennary glycans
\item \code{A3Fc0B}: Proportion of bisecting glycans within a-core-fucosylated tri-antennary glycans
\item \code{A4Fc0B}: Proportion of bisecting glycans within a-core-fucosylated tetra-antennary glycans
\item \code{A1G}: Average degree of galactosylation per antenna within mono-antennary glycans
\item \code{A2G}: Average degree of galactosylation per antenna within bi-antennary glycans
\item \code{A3G}: Average degree of galactosylation per antenna within tri-antennary glycans
\item \code{A4G}: Average degree of galactosylation per antenna within tetra-antennary glycans
\item \code{A1Gt}: Average degree of terminal galactose per antenna within mono-antennary glycans
\item \code{A2Gt}: Average degree of terminal galactose per antenna within bi-antennary glycans
\item \code{A3Gt}: Average degree of terminal galactose per antenna within tri-antennary glycans
\item \code{A4Gt}: Average degree of terminal galactose per antenna within tetra-antennary glycans
\item \code{A1S}: Average degree of sialylation per antenna within mono-antennary glycans
\item \code{A2S}: Average degree of sialylation per antenna within bi-antennary glycans
\item \code{A3S}: Average degree of sialylation per antenna within tri-antennary glycans
\item \code{A4S}: Average degree of sialylation per antenna within tetra-antennary glycans
\item \code{A1GS}: Average degree of sialylation per galactose within mono-antennary glycans
\item \code{A2GS}: Average degree of sialylation per galactose within bi-antennary glycans
\item \code{A3GS}: Average degree of sialylation per galactose within tri-antennary glycans
\item \code{A4GS}: Average degree of sialylation per galactose within tetra-antennary glycans
}

These additional sialic acid linkage traits are included if \code{sia_link = TRUE}.
\itemize{
\item \code{A1E}: Average degree of a2,6-linked sialylation per antenna within mono-antennary glycans
\item \code{A2E}: Average degree of a2,6-linked sialylation per antenna within bi-antennary glycans
\item \code{A3E}: Average degree of a2,6-linked sialylation per antenna within tri-antennary glycans
\item \code{A4E}: Average degree of a2,6-linked sialylation per antenna within tetra-antennary glycans
\item \code{A1L}: Average degree of a2,3-linked sialylation per antenna within mono-antennary glycans
\item \code{A2L}: Average degree of a2,3-linked sialylation per antenna within bi-antennary glycans
\item \code{A3L}: Average degree of a2,3-linked sialylation per antenna within tri-antennary glycans
\item \code{A4L}: Average degree of a2,3-linked sialylation per antenna within tetra-antennary glycans
\item \code{A1GE}: Average degree of a2,6-linked sialylation per galactose within mono-antennary glycans
\item \code{A2GE}: Average degree of a2,6-linked sialylation per galactose within bi-antennary glycans
\item \code{A3GE}: Average degree of a2,6-linked sialylation per galactose within tri-antennary glycans
\item \code{A4GE}: Average degree of a2,6-linked sialylation per galactose within tetra-antennary glycans
\item \code{A1GL}: Average degree of a2,3-linked sialylation per galactose within mono-antennary glycans
\item \code{A2GL}: Average degree of a2,3-linked sialylation per galactose within bi-antennary glycans
\item \code{A3GL}: Average degree of a2,3-linked sialylation per galactose within tri-antennary glycans
\item \code{A4GL}: Average degree of a2,3-linked sialylation per galactose within tetra-antennary glycans
}
}
\section{Usage of sialic acid linkage traits}{
To use these sialic acid linkage traits,
\code{rowData()} of the input \code{GlycomicSE} or \code{GlycoproteomicSE} must have the
following columns:
\itemize{
\item \code{nE}: Number of a2,6-linked sialic acids
\item \code{nL}: Number of a2,3-linked sialic acids
}

Note that you have to add these two columns even if the \code{glycan_structure} column has intact linkages.
This is because by convention all traits work with glycan structures with "basic" structure levels
(i.e., with generic monosaccharides like "Hex" and "HexNAc" and no linkages specified).
}

\examples{
traits_detailed()

}
