% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{traits_fu_2026}
\alias{traits_fu_2026}
\title{Get Traits in Fu et al. 2026}
\usage{
traits_fu_2026()
}
\value{
A named list of derived traits.
}
\description{
These traits are the ones used by Fu et al. 2026 (https://doi.org/10.1038/s41467-026-68579-x).
It is much like \code{all_traits()}, but doesn't differentiate core- and arm-fucosylation,
and doesn't include the sialic acid linkage traits.
Also, many traits are too specific to be useful and interpretable.
}
\examples{
traits_fu_2026()[1:5]
}
