% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{traits_li_2025}
\alias{traits_li_2025}
\title{Get Traits in Li et al. 2025}
\usage{
traits_li_2025()
}
\value{
A named list of derived traits.
}
\description{
These traits are the ones used by Li et al. 2025 (https://doi.org/10.1038/s41467-025-57633-9).
These traits were created based only on glycan compositions, not structures.
Here the traits are remapped to use glycan structures,
so they are not exactly the same as the ones in the paper.
Generally we don't recommend using these traits because they capture too little information.
We include this function because this paper is very influential in the field.
}
\examples{
traits_li_2025()[1:5]
}
