% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-factory.R
\name{wmean}
\alias{wmean}
\title{Create a Weighted-Mean Trait}
\usage{
wmean(val_cond, within = NULL, na_action = "keep")
}
\arguments{
\item{val_cond}{Condition to use for defining the value.
An expression that evaluates to a numeric vector.
The names of all built-in meta-properties (see \code{\link[=all_mp_fns]{all_mp_fns()}}) and custom meta-properties
can be used in the expression.}

\item{within}{Condition to set a restriction for the glycans. Same format as \code{val_cond}.}

\item{na_action}{How to handle missing values.
\itemize{
\item "keep" (default): keep the missing values as NA.
\item "zero": set the missing values to 0.
}}
}
\value{
A derived trait function.
}
\description{
A weighted-mean trait is the average value of some quantitative property within a group of glycans,
weighted by the abundance of the glycans.
For example, the average number of antennae within all complex glycans,
or the average number of sialic acids within all glycans.
}
\details{
You can use \code{wmean()} to create weighted-mean trait easily.

For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Weighted mean of the number of sialic acids within all glycans
wmean(nS)

# Average degree of sialylation per antenna within all glycans
wmean(nS / nA)
}\if{html}{\out{</div>}}

Note that the last example uses \code{/} for division.
Actually, you can use any arithmetic operator in the expression in R (e.g., \code{*}, \code{+}, \code{-}, etc.).

If you want to perform a pre-filtering before calculating the weighted-mean,
for example, you want to calculate the average degree of sialylation per antenna within only complex glycans,
you can use \code{within} to define the restriction.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Average number of antennae within complex glycans
wmean(nA, within = (Tp == "complex"))
}\if{html}{\out{</div>}}
}
\examples{
# Weighted mean of the number of sialic acids within all glycans
wmean(nS)

# Average degree of sialylation per antenna within all glycans
wmean(nS / nA)

# Average number of antennae within complex glycans
wmean(nA, within = (Tp == "complex"))

}
