% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-factory.R
\name{wsum}
\alias{wsum}
\title{Create a Weighted Sum Trait}
\usage{
wsum(val, within = NULL)
}
\arguments{
\item{val}{Expression to use for defining the value.
An expression that evaluates to a logical vector.
The names of all built-in meta-properties (see \code{\link[=all_mp_fns]{all_mp_fns()}}) and custom meta-properties
can be used in the expression.}

\item{within}{Condition to set a restriction for the glycans. Same format as \code{val}.}
}
\value{
A derived trait function.
}
\description{
A weighted sum trait is the sum of a quantitative property within a group of glycans,
weighted by the abundance of the glycans.
For example, the sum of the number of sialic acids within all glycans,
or the sum of the number of Lewis x antigens within all glycans.
}
\section{How to use}{


You can use \code{wsum()} to create weighted sum trait easily.

For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Weighted sum of the number of sialic acids within all glycans
wsum(nS)
}\if{html}{\out{</div>}}

This can be regarded as the quantification of sialic acids.
If some glycan has only one sialic acid,
its abundance is added to the results.
If another glycan has two sialic acids,
its abundance is doubled before being added to the results.

You can also use \code{within} to restrict the weighted sum calculation to specific glycan subsets.
For example, you can calculate the weighted sum of the number of sialic acids within complex glycans:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{wsum(nS, within = (Tp == "complex"))
}\if{html}{\out{</div>}}
}

\examples{
# Weighted sum of the number of sialic acids within all glycans
wsum(nS)

# Weighted sum of the number of sialic acids within complex glycans
wsum(nS, within = (Tp == "complex"))

}
