% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-traits.R
\name{all_traits}
\alias{all_traits}
\title{Get All Derived Traits}
\usage{
all_traits()
}
\value{
A named list of derived traits.
}
\description{
This function returns a named list of all derived traits:
\itemize{
\item "TM": \code{prop(T == "highmannose")}
\item "TH": \code{prop(T == "hybrid")}
\item "TC": \code{prop(T == "complex")}
\item "MM": \code{wmean(nM, within = (T == "highmannose"))}
\item "CA2": \code{prop(nA == 2, within = (T == "complex"))}
\item "CA3": \code{prop(nA == 3, within = (T == "complex"))}
\item "CA4": \code{prop(nA == 4, within = (T == "complex"))}
\item "TF": \code{prop((nFc + nFa) > 0)}
\item "TFc": \code{prop(nFc > 0)}
\item "TFa": \code{prop(nFa > 0)}
\item "TB": \code{prop(B)}
\item "SG": \code{wmean(nS / nG)}
\item "GA": \code{wmean(nG / nA)}
\item "TS": \code{prop(nS > 0)}
}
}
