% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glydet-package.R
\docType{package}
\name{glydet-package}
\alias{glydet}
\alias{glydet-package}
\title{glydet: Describe Glycosylation Structural Properties in a Site Specific Manner}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Describe glycosylation structural properties in a site specific manner. In the field of glycomics, this analytical approach is known as derived traits. glydet provides functions to calculate derived traits well-defined in literature, and implements a domain-specific language to define custom derived traits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glydet/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
