% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-trait.R
\name{make_trait}
\alias{make_trait}
\title{Use a Large Language Model (LLM) to create a derived trait function}
\usage{
make_trait(description)
}
\arguments{
\item{description}{A description of the trait in natural language.}
}
\value{
A derived trait function.
}
\description{
This function allows you to create a derived trait function using natural language.
Note that LLMs can be unreliable, so the result should be verified manually.
If the description is not clear, an error will be raised.
Try to read the descriptions of built-in traits to get ideas.
Currently, only \code{prop()}, \code{ratio()}, and \code{wmean()} are supported.
To use this feature, you need to install the \code{ellmer} package.
You also need to provide an API key for the DeepSeek chat model.
Please set the environment variable \code{DEEPSEEK_API_KEY} to your API key using \code{Sys.setenv()}.
You can obtain an API key from https://platform.deepseek.com.
}
\examples{
# Sys.setenv(DEEPSEEK_API_KEY = "your_api_key")
# make_trait("the average number of sialic acids")

}
